move pacakge exact to start script
