+++ scripts/start.js
@@ -6,5 +6,8 @@ require('./utils/env')
 // Ensure dependency requirements are met.
 require('./utils/checkDependencies')
 
+//Ensure only have exact versions installed
+require('./utils/package-exact')
+
 // expose the app
 module.exports = require('../index.js')
